fix(windsurf): update shell scripts

+++ windsurf/apply.sh
@@ -2,4 +2,7 @@
 BASEDIR="$(dirname "$(realpath "$0")")"
 SETTING_PATH="$HOME/Library/Application Support/Windsurf/User/settings.json"
 
-cp "$BASEDIR/settings.json" "$SETTING_PATH"
+if [ -f "$SETTING_PATH" ]; then
+  echo "Updating Windsurf setting"
+  cp "$BASEDIR/settings.json" "$SETTING_PATH"
+fi
